extract logging adapters and disable notification and shortcut for release by default
