+++ plugins/quick-brick-xray/android/src/main/java/com/applicaster/plugin/xray/XRayPlugin.kt
@@ -8,14 +8,15 @@ import android.content.pm.ShortcutInfo
 import android.content.pm.ShortcutManager
 import android.graphics.drawable.Icon
 import android.os.Build
-import android.util.Log
-import android.util.Printer
+import com.applicaster.plugin.xray.logadapters.APLoggerAdapter
+import com.applicaster.plugin.xray.logadapters.FLogAdapter
+import com.applicaster.plugin.xray.logadapters.PrinterAdapter
 import com.applicaster.plugin.xray.ui.LogActivity
 import com.applicaster.plugin_manager.crashlog.CrashlogPlugin
+import com.applicaster.util.APDebugUtil
 import com.applicaster.util.APLogger
 import com.applicaster.util.AppContext
 import com.applicaster.util.StringUtil
-import com.applicaster.util.logging.IAPLogger
 import com.applicaster.xray.android.routing.DefaultSinkFilter
 import com.applicaster.xray.android.sinks.ADBSink
 import com.applicaster.xray.android.sinks.PackageFileLogSink
@@ -27,8 +28,6 @@ import com.applicaster.xray.crashreporter.SendActivity
 import com.applicaster.xray.example.sinks.InMemoryLogSink
 import com.applicaster.xray.ui.notification.XRayNotification
 import com.facebook.common.logging.FLog
-import com.facebook.common.logging.LoggingDelegate
-import com.facebook.debug.debugoverlay.model.DebugOverlayTag
 import com.facebook.debug.holder.NoopPrinter
 import com.facebook.debug.holder.PrinterHolder
 
@@ -79,120 +78,17 @@ class XRayPlugin : CrashlogPlugin {
     }
 
     private fun overrideRNPrinter() {
-        PrinterHolder.setPrinter(object : Printer, com.facebook.debug.holder.Printer {
-
-            val logger = Logger.get("ReactNative/DebugPrinter")
-
-            override fun println(x: String?) {
-                logger.d(TAG).message(x!!)
-            }
-
-            override fun logMessage(tag: DebugOverlayTag?, message: String?, vararg args: Any?) {
-                logger.d(tag!!.name).message(java.lang.String.format(message!!, *args))
-            }
-
-            override fun logMessage(tag: DebugOverlayTag?, message: String?) {
-                logger.d(tag!!.name).message(message!!)
-            }
-
-            override fun shouldDisplayLogMessage(tag: DebugOverlayTag?): Boolean {
-                return true
-            }
-
-        })
+        PrinterHolder.setPrinter(PrinterAdapter())
         pluginLogger.i(TAG).message("React native printer is now intercepted by X-Ray")
     }
 
     private fun hookRNLogger() {
-        FLog.setLoggingDelegate(object : LoggingDelegate {
-
-            val logger = Logger.get("ReactNative")
-            var level = Log.DEBUG // log at debug level since we have our own filters
-
-            override fun wtf(tag: String?, msg: String?) {
-                this.logger.e(tag!!).message(msg!!)
-            }
-
-            override fun wtf(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.e(tag!!).exception(tr!!).message(msg!!)
-            }
-
-            override fun getMinimumLoggingLevel(): Int = level
-
-            override fun w(tag: String?, msg: String?) {
-                this.logger.w(tag!!).message(msg!!)
-            }
-
-            override fun w(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.w(tag!!).exception(tr!!).message(msg!!)
-            }
-
-            override fun v(tag: String?, msg: String?) {
-                this.logger.v(tag!!).message(msg!!)
-            }
-
-            override fun v(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.v(tag!!).exception(tr!!).message(msg!!)
-            }
-
-            override fun log(priority: Int, tag: String?, msg: String?) {
-                this.logger.e(tag!!).message(msg!!)
-            }
-
-            override fun setMinimumLoggingLevel(level: Int) {
-                this.level = level
-            }
-
-            override fun isLoggable(level: Int): Boolean {
-                return this.level >= level
-            }
-
-            override fun i(tag: String?, msg: String?) {
-                this.logger.i(tag!!).message(msg!!)
-            }
-
-            override fun i(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.i(tag!!).exception(tr!!).message(msg!!)
-            }
-
-            override fun e(tag: String?, msg: String?) {
-                this.logger.e(tag!!).message(msg!!)
-            }
-
-            override fun e(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.e(tag!!).exception(tr!!).message(msg!!)
-            }
-
-            override fun d(tag: String?, msg: String?) {
-                this.logger.d(tag!!).message(msg!!)
-            }
-
-            override fun d(tag: String?, msg: String?, tr: Throwable?) {
-                this.logger.d(tag!!).exception(tr!!).message(msg!!)
-            }
-
-        })
+        FLog.setLoggingDelegate(FLogAdapter())
         pluginLogger.i(TAG).message("React native logger is now intercepted by X-Ray")
     }
 
     private fun hookApplicasterLogger() {
-        APLogger.setLogger(object : IAPLogger {
-
-            private val logger = Logger.get("ApplicasterSDK")
-
-            override fun verbose(tag: String, msg: String) = logger.v(tag).message(msg)
-
-            override fun debug(tag: String, msg: String) = logger.d(tag).message(msg)
-
-            override fun info(tag: String, msg: String) = logger.i(tag).message(msg)
-
-            override fun warn(tag: String, msg: String) = logger.w(tag).message(msg)
-
-            override fun error(tag: String, msg: String) = logger.e(tag).message(msg)
-
-            override fun error(tag: String, msg: String, t: Throwable) =
-                    logger.e(tag).exception(t).message(msg)
-        })
+        APLogger.setLogger(APLoggerAdapter())
         pluginLogger.i(TAG).message("Applicaster APLogger is now intercepted by X-Ray")
     }
 
@@ -215,10 +111,10 @@ class XRayPlugin : CrashlogPlugin {
                     .addSink(fileSinkKey, fileSink)
                     .setFilter(fileSinkKey, "", DefaultSinkFilter(eFileLogLevel))
             // enable our own crash reports sending, but do not handle crashes
-            Reporting.init(reportEmail?:"", fileSink.file)
+            Reporting.init(reportEmail ?: "", fileSink.file)
         } else {
             // enable basic reporting without file (not very useful)
-            Reporting.init(reportEmail?:"", null)
+            Reporting.init(reportEmail ?: "", null)
         }
 
         // todo: we need immediate crash reporting mode in zapp: show share intent right after the crash
@@ -229,33 +125,11 @@ class XRayPlugin : CrashlogPlugin {
             // can't disable it right now, since previous exception handler is lost, and this code could be called twice
         }
 
-        val showNotification = StringUtil.booleanValue(configuration?.get(notificationKey))
+        val showNotification = APDebugUtil.getIsInDebugMode()
+                && StringUtil.booleanValue(configuration?.get(notificationKey))
 
         if(showNotification) {
-            val showLogIntent = PendingIntent.getActivity(
-                    context,
-                    0,
-                    Intent(context, LogActivity::class.java)
-                            .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK),
-                    PendingIntent.FLAG_CANCEL_CURRENT
-            )!!
-
-            // actions order is kept in the UI
-            val actions: HashMap<String, PendingIntent> = linkedMapOf("Show" to showLogIntent)
-
-            if(fileLogLevel != null) {
-                // add report sharing button to notification
-                // if we have file logging enabled, allow to send it
-                val shareLogIntent = SendActivity.getSendPendingIntent(context)
-                actions.put("Send", shareLogIntent)
-            }
-
-            // here we show Notification UI with custom actions
-            XRayNotification.show(
-                    context,
-                    101,
-                    actions
-            )
+            setupNotification(fileLogLevel)
         }
 
         val logRNPrinter =  StringUtil.booleanValue(configuration?.get(debugRNKey))
@@ -266,20 +140,55 @@ class XRayPlugin : CrashlogPlugin {
         }
 
         // add shortcut
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
-            context.getSystemService<ShortcutManager>(ShortcutManager::class.java)?.let { shortcutManager ->
-                if (!shortcutManager.dynamicShortcuts.stream().anyMatch { it.id == "xray" }) {
-                    val shortcut = ShortcutInfo.Builder(context, "xray")
-                            .setShortLabel("XRay")
-                            .setLongLabel("Open XRay log")
-                            .setIcon(Icon.createWithResource(context, R.drawable.ic_xray_settings_24))
-                            .setIntent(Intent(context, LogActivity::class.java).setAction(Intent.ACTION_VIEW))
-                            .build()
-                    shortcutManager.addDynamicShortcuts(listOf(shortcut))
-                }
+        if (APDebugUtil.getIsInDebugMode()) {
+            setupShortcut()
+        }
+
+        // todo: update filters configuration
+    }
+
+    private fun setupShortcut() {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N_MR1) {
+            return
+        }
+        context.getSystemService<ShortcutManager>(ShortcutManager::class.java)?.let { shortcutManager ->
+            if (!shortcutManager.dynamicShortcuts.stream().anyMatch { it.id == "xray" }) {
+                val shortcut = ShortcutInfo.Builder(context, "xray")
+                        .setShortLabel("XRay")
+                        .setLongLabel("Open XRay log")
+                        .setIcon(Icon.createWithResource(context, R.drawable.ic_xray_settings_24))
+                        .setIntent(Intent(context, LogActivity::class.java).setAction(Intent.ACTION_VIEW))
+                        .build()
+                shortcutManager.addDynamicShortcuts(listOf(shortcut))
             }
         }
-        // todo: update filters configuration
+    }
+
+    private fun setupNotification(fileLogLevel: String?) {
+        val showLogIntent = PendingIntent.getActivity(
+                context,
+                0,
+                Intent(context, LogActivity::class.java)
+                        .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK),
+                PendingIntent.FLAG_CANCEL_CURRENT
+        )!!
+
+        // actions order is kept in the UI
+        val actions: HashMap<String, PendingIntent> = linkedMapOf("Show" to showLogIntent)
+
+        if (fileLogLevel != null) {
+            // add report sharing button to notification
+            // if we have file logging enabled, allow to send it
+            val shareLogIntent = SendActivity.getSendPendingIntent(context)
+            actions.put("Send", shareLogIntent)
+        }
+
+        // here we show Notification UI with custom actions
+        XRayNotification.show(
+                context,
+                101,
+                actions
+        )
     }
 
 }
